feat: update to go 1.27
