+++ starknet-p2p-specs/go.mod
@@ -1,5 +1,5 @@
 module github.com/starknet-io/starknet-p2p-specs
 
-go 1.26.0
+go 1.27.0
 
 require google.golang.org/protobuf v1.36.11
